requirements.txt
Flask
tensorflow==2.13
numpy
pillow
gunicorn